Add New Startup File Version Response key value
