+++ NetCoreProj/Program.cs
@@ -20,7 +20,7 @@ namespace NetCoreProj
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
-                    webBuilder.UseStartup<Startup>();
+                    webBuilder.UseStartup<MyStartup>();
                 });
     }
 }
